rename folder to ReportSuggestionForm
